LayoutTests/fast/dom/Element/attribute-setAttributeNode-multiple-times-expected.txt

Test using setAttributeNode() multiple times with "compatible" names and "incompatible" names."

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing: FooBar
PASS document.documentElement.attributes.length is 1
PASS document.documentElement.attributes.length is 0
PASS document.documentElement.attributes.length is 1
PASS document.documentElement.attributes.length is 0

Testing: foobar
PASS document.documentElement.attributes.length is 1
PASS document.documentElement.attributes.length is 0
PASS document.documentElement.attributes.length is 1
PASS document.documentElement.attributes.length is 0

Testing: FOOBAR
PASS document.documentElement.attributes.length is 1
PASS document.documentElement.attributes.length is 0
PASS document.documentElement.attributes.length is 1
PASS document.documentElement.attributes.length is 0

PASS successfullyParsed is true

TEST COMPLETE
